+++ updates/multicast/multicast_region_result.json
@@ -1319,8 +1319,7 @@
                 "rtsp://112.86.202.37:554/JSBC_iptv/C10000009@JSBC"
             ],
             "电信": [
-                "rtp://239.49.8.75:8000",
-                "rtp://239.49.8.18:9610"
+                "rtp://239.49.8.75:8000"
             ]
         },
         "江西": {
@@ -4581,6 +4580,9 @@
             "联通": [
                 "rtsp://112.86.202.37:554/JSBC_iptv/C10000040@JSBC",
                 "rtsp://157.0.143.37:554/JSBC_iptv/C10000040@JSBC"
+            ],
+            "电信": [
+                "rtp://239.49.8.18:9610"
             ]
         },
         "吉林": {
@@ -13215,6 +13217,11 @@
                 "rtp://239.1.0.175:1025"
             ]
         },
+        "江苏": {
+            "电信": [
+                "rtp://239.49.8.31:8000"
+            ]
+        },
         "吉林": {
             "电信": [
                 "rtp://239.37.1.21:5540"
@@ -14426,6 +14433,11 @@
                 "rtp://239.76.252.254:9000"
             ]
         },
+        "江苏": {
+            "电信": [
+                "rtp://239.49.0.125:8000"
+            ]
+        },
         "辽宁": {
             "电信": [
                 "rtp://239.33.4.220:22200"
@@ -15331,6 +15343,11 @@
         "湖南": {
             "移动": [
                 "rtp://239.1.1.39:1025"
+            ]
+        },
+        "江苏": {
+            "电信": [
+                "rtp://239.49.0.111:8000"
             ]
         },
         "江西": {
@@ -22349,10 +22366,25 @@
             ]
         }
     },
-    "凤凰資訊": {
+    "凤凰资讯": {
         "福建": {
             "移动": [
                 "http://183.251.61.207/PLTV/88888888/224/3221225901/index.m3u8"
+            ]
+        },
+        "甘肃": {
+            "电信": [
+                "rtp://239.255.30.70:8231"
+            ]
+        },
+        "湖南": {
+            "电信": [
+                "rtp://239.76.246.8:1234"
+            ]
+        },
+        "陕西": {
+            "电信": [
+                "rtp://239.111.205.82:5140"
             ]
         }
     },
@@ -22466,6 +22498,9 @@
             "联通": [
                 "rtsp://112.86.202.37:554/JSBC_iptv/C10000228@JSBC",
                 "rtsp://157.0.143.37:554/JSBC_iptv/C10000228@JSBC"
+            ],
+            "电信": [
+                "rtp://239.49.8.119:8000"
             ]
         },
         "辽宁": {
@@ -24023,23 +24058,6 @@
         "甘肃": {
             "电信": [
                 "rtp://239.255.30.252:8231"
-            ]
-        }
-    },
-    "凤凰资讯": {
-        "甘肃": {
-            "电信": [
-                "rtp://239.255.30.70:8231"
-            ]
-        },
-        "湖南": {
-            "电信": [
-                "rtp://239.76.246.8:1234"
-            ]
-        },
-        "陕西": {
-            "电信": [
-                "rtp://239.111.205.82:5140"
             ]
         }
     },
@@ -34420,7 +34438,7 @@
             ]
         }
     },
-    "吉林乾安综合普清": {
+    "吉林干安综合普清": {
         "吉林": {
             "电信": [
                 "rtp://239.37.0.181:5540"
